+++ MongoDB.Entities/Relationships/Many.Fluent.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using MongoDB.Bson;
 using MongoDB.Driver;
 
@@ -101,11 +102,13 @@ public sealed partial class Many<TChild, TParent> where TChild : IEntity where T
     /// <param name="options">An optional AggregateOptions object</param>
     public IAggregateFluent<TParent> ParentsFluent(IEnumerable<object> childIDs, IClientSessionHandle? session = null, AggregateOptions? options = null)
     {
+        var childIds = childIDs.Select(Cache<TChild>.IdToBsonValue).ToArray();
+
         return typeof(TParent) == typeof(TChild)
                    ? throw new InvalidOperationException("Both parent and child types cannot be the same")
                    : _isInverse
                        ? JoinFluent(session, options)
-                         .Match(f => f.In(j => j.ParentID, childIDs))
+                         .Match(f => f.In(j => j.ParentID, childIds))
                          .Lookup<JoinRecord, TParent, Joined<TParent>>(
                              _db.Collection<TParent>(),
                              j => j.ChildID,
@@ -114,7 +117,7 @@ public sealed partial class Many<TChild, TParent> where TChild : IEntity where T
                          .ReplaceRoot(j => j.Results[0])
                          .Distinct()
                        : JoinFluent(session, options)
-                         .Match(f => f.In(j => j.ChildID, childIDs))
+                         .Match(f => f.In(j => j.ChildID, childIds))
                          .Lookup<JoinRecord, TParent, Joined<TParent>>(
                              _db.Collection<TParent>(),
                              r => r.ParentID,
@@ -133,9 +136,11 @@ public sealed partial class Many<TChild, TParent> where TChild : IEntity where T
     {
         _parent.ThrowIfUnsaved();
 
+        var parentId = _parent.GetBsonId();
+
         return _isInverse
                    ? JoinFluent(session, options)
-                     .Match(f => f.Eq(r => r.ChildID, _parent.GetId()))
+                     .Match(f => f.Eq(r => r.ChildID, parentId))
                      .Lookup<JoinRecord, TChild, Joined<TChild>>(
                          _db.Collection<TChild>(),
                          r => r.ParentID,
@@ -143,7 +148,7 @@ public sealed partial class Many<TChild, TParent> where TChild : IEntity where T
                          j => j.Results)
                      .ReplaceRoot(j => j.Results[0])
                    : JoinFluent(session, options)
-                     .Match(f => f.Eq(r => r.ParentID, _parent.GetId()))
+                     .Match(f => f.Eq(r => r.ParentID, parentId))
                      .Lookup<JoinRecord, TChild, Joined<TChild>>(
                          _db.Collection<TChild>(),
                          r => r.ChildID,