+++ Tests/Models/Flower/FlowerEntity.cs
@@ -7,7 +7,7 @@ namespace MongoDB.Entities.Tests;
 [Collection("FlowerEntity")]
 public class FlowerEntity : Flower
 {
-    [BsonId, AsObjectId]
+    [BsonId]
     public string Id { get; set; }
 
     public override object GenerateNewID()