+++ Tests/Models/Place/PlaceEntity.cs
@@ -6,7 +6,7 @@ namespace MongoDB.Entities.Tests;
 [Collection("PlaceEntity")]
 public class PlaceEntity : Place
 {
-    [BsonId, AsObjectId]
+    [BsonId]
     public string? Id { get; set; }
 
     public override object GenerateNewID()