+++ Tests/Models/Review/ReviewEntity.cs
@@ -7,7 +7,7 @@ namespace MongoDB.Entities.Tests;
 [Collection("ReviewEntity")]
 public class ReviewEntity : Review
 {
-    [BsonId, AsObjectId]
+    [BsonId]
     public string Id { get; set; }
 
     public override object GenerateNewID()